install/deps: Factor optional failure reporting into its own module
PR-URL: https://github.com/npm/npm/pull/13692/
Credit: @iarna
