+++ lib/install/deps.js
@@ -30,6 +30,7 @@ var packageId = require('../utils/package-id.js')
 var moduleName = require('../utils/module-name.js')
 var isDevDep = require('./is-dev-dep.js')
 var isProdDep = require('./is-prod-dep.js')
+var reportOptionalFailure = require('./report-optional-failure.js')
 
 // The export functions in this module mutate a dependency tree, adding
 // items to them.
@@ -349,18 +350,6 @@ var failedDependency = exports.failedDependency = function (tree, name_pkg) {
   return false
 }
 
-function top (tree) {
-  if (tree.parent) return top(tree.parent)
-  return tree
-}
-
-function treeWarn (tree, what, error) {
-  var topTree = top(tree)
-  if (!topTree.warnings) topTree.warnings = []
-  error.optional = flatNameFromTree(tree) + '/' + what
-  topTree.warnings.push(error)
-}
-
 function andHandleOptionalErrors (log, tree, name, done) {
   validate('OOSF', arguments)
   return function (er, child, childLog) {
@@ -369,7 +358,7 @@ function andHandleOptionalErrors (log, tree, name, done) {
     var isFatal = failedDependency(tree, name)
     if (er && !isFatal) {
       tree.children = tree.children.filter(noModuleNameMatches(name))
-      treeWarn(tree, name, er)
+      reportOptionalFailure(tree, name, er)
       return done()
     } else {
       return done(er, child, childLog)
